[refactor] #58 LoginUser 패키지 위치 변경
