+++ src/main/java/org/lostnomore/backend/global/resolver/LoginUser.java
@@ -1,4 +1,4 @@
-package org.lostnomore.backend.global;
+package org.lostnomore.backend.global.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
